services/genworker-app/requirements.txt
python-socketio[client]
python-socketio
aiortc
rich
diffusers
torch
watchfiles
aiohttp
transformers
accelerate